Fixed a crash when resetting map cache before folder exists

+++ src/ClassicUO.Client/Game/UI/Gumps/WorldMapGump.cs
@@ -404,7 +404,8 @@ public class WorldMapGump : ResizableGump
 
         _options["reset_map_cache"] = new ContextMenuItemEntry(ResGumps.ResetMapsCache, () =>
         {
-            Directory.GetFiles(_mapsCachePath, "*.png").ForEach(s => File.Delete(s));
+            if(Directory.Exists(_mapsCachePath))
+                Directory.GetFiles(_mapsCachePath, "*.png").ForEach(s => File.Delete(s));
         }, false);
     }
